requirements.txt
python-telegram-bot==13.10
boto==2.49.0
finviz==1.4.3
requests==2.27.1
pytz==2021.3
beautifulsoup4==4.10.0
pandas==1.4.0
dataframe-image==0.1.1